made voting buttons a component
